requirements.txt
langchain==0.0.311
python-dotenv==1.0.0
openai==0.28.1
faiss-cpu==1.7.4
tiktoken==0.5.1
pypdf==3.17.1
streamlit==1.29.0
